# api/swagger/swagger.yaml
---
info:
  description: Demo de una Rest Api
  title: Demo Rest API
  version: 1.0.0
host: localhost:3000
basePath: '/'
produces:
- application/json
schemes:
- http
swagger: '2.0'
paths:
  /login:
    post:
      description: Funcion que loguea a un usuario con la BD
      tags:
        - 'auth '
      parameters: 
        - in: body
          name: usuario
          description: El email del usuario a loguearse
          schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
          responses:
            200:
              description: Ok
  /registrar:
    post:
      description: Funcion que registra a un usuario en la BD
      tags:
        - 'auth '
      parameters: 
        - in: body
          name: usuario
          description: El email del usuario a loguearse
          schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
          responses:
            201:
              description: Created
  /articles:
    get:
      parameters: []
      security:
        - APIKeyHeader: []
      description: Esta funcion devuelve articulos obtenidos desde la base de datos
      tags:
      - 'articulos '
      responses:
        200:
          description: "Lista de articulos"
        401:
          description: API key is missing or invalid
    post:
      parameters: []
      description: Esta funcion crea un articulo
      tags:
      - 'articulos '
      responses:
        201:
          description: " {}"
  /articles/{id}:
    get:
      parameters:
      - name: id
        in: path
        description: id
        required: true
        type: number
      description: Este Recurso devuelve articulos obtenidos desde la base de datos
      tags:
      - 'articulos '
      responses:
        200:
          description: " {}"
    put:
      parameters: 
      - name: id
        in: path
        description: id
        required: true
        type: number
      description: Recurso para modificar un Articulo
      tags:
      - 'articulos '
      responses:
        201:
          description: " {}"
    delete:
      parameters: 
      - name: id
        in: path
        description: id
        required: true
        type: number
      description: Recurso para eliminar un Articulo
      tags:
      - 'articulos '
      responses:
        202:
          description: " {}"
  /posts:
    get:
      parameters: []
      description: Esta funcion devuelve la lista de posts
      tags:
      - 'posts '
      responses:
        200:
          description: " Lista de articulos"
    post:
      parameters: []
      description: Esta funcion crea un post
      tags:
      - 'posts '
      responses:
        201:
          description: " {}"
  /posts/{id}:
    get:
      parameters:
      - name: id
        in: path
        description: id
        required: true
        type: number
      description: Esta funcion devuelve un post especifico filtrado por id
      tags:
      - 'posts '
      responses:
        200:
          description: " post"
    put:
      parameters: 
      - name: id
        in: path
        description: id
        required: true
        type: number
      description: Esta funcion modifica un post
      tags:
      - 'posts '
      responses:
        201:
          description: " {}"
    patch:
      parameters: 
      - name: id
        in: path
        description: id
        required: true
        type: number
      description: Esta funcion modifica un atributos especificos de un post
      tags:
      - 'posts '
      responses:
        201:
          description: " {}"
    delete:
      parameters: 
      - name: id
        in: path
        description: id
        required: true
        type: number
      description: Esta funcion borra un post
      tags:
      - 'posts '
      responses:
        202:
          description: " {}"
definitions: {}
responses: {}
parameters: {}
securityDefinitions: 
  APIKeyHeader:
     type: apiKey
     in: header
     name: X-API-Key
tags:
- name: 'articulos '
  description: " Operacion con articulos"
- name: 'posts '
  description: " Operaciones con Posts"
- name: 'auth '
  description: " Operaciones de Authenticacion"